pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "youtube-analyzer"
version = "0.1.0"
description = "YouTube Network Analytics GUI"
authors = [
    {name = "Ross",email = "ross.kugler@wsu.edu"}
]
readme = "README.md"
requires-python = "3.11.9"
dependencies = [
    "streamlit (>=1.51.0,<2.0.0)",
    "pyspark (==3.5.3)",
    "pymongo (>=4.15.4,<5.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "numpy (>=2.3.5,<3.0.0)",
    "plotly (>=6.5.0,<7.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "networkx (>=3.5,<4.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
